clean Upload file
