--- uikit/domql/IconTextWithNotification/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@symbo.ls/icontextwithnotification",
-  "version": "2.11.132",
-  "main": "index.js",
-  "license": "MIT",
-  "gitHead": "dfc71a41e1d64da4fef5ad0ecc0b8f6a3be79890",
-  "dependencies": {
-    "@symbo.ls/atoms": "latest",
-    "@symbo.ls/utils": "latest"
-  },
-  "source": "src/index.js"
-}
