+++ uikit/domql/Progress/index.js
@@ -4,13 +4,6 @@ import { Flex } from '@symbo.ls/atoms'
 
 export const ProgressLine = {
   tag: 'progress',
-
-  attr: {
-    max: ({ props }) => props.max,
-    progress: ({ props }) => props.progress,
-    value: ({ props }) => props.value
-  },
-
   props: {
     value: 0.7,
     height: 'Y+V',
@@ -25,30 +18,26 @@ export const ProgressLine = {
         borderRadius: '2px'
       }
     }
+  },
+  attr: {
+    max: ({ props }) => props.max,
+    progress: ({ props }) => props.progress,
+    value: ({ props }) => props.value
   }
 }
 
 export const ProgressLineWithUnitValue = {
   extend: Flex,
-
-  ProgressLine: {},
-  UnitValue: {},
-
   props: {
     align: 'center flex-start',
     gap: 'A'
-  }
+  },
+
+  ProgressLine: {},
+  UnitValue: {}
 }
 
 export const ProgressCircle = {
-  progress: {
-    attr: {
-      max: ({ props }) => props.max,
-      progress: ({ props }) => props.progress,
-      value: ({ props }) => props.value
-    }
-  },
-
   props: {
     boxSize: 'D D',
     position: 'relative',
@@ -62,9 +51,12 @@ export const ProgressCircle = {
       left: '50%',
       transform: 'translate(-50%, -50%)',
       round: '100%'
-    },
+    }
+  },
 
-    progress: {
+  Progress: {
+    tag: 'progress',
+    props: {
       boxSize: '100% 100%',
       value: 0.7,
       round: '100%',
@@ -73,59 +65,55 @@ export const ProgressCircle = {
         '&::-webkit-progress-bar': { background: '#0474F2' },
         '&::-webkit-progress-value': { background: '#252527' }
       }
+    },
+    attr: {
+      max: ({ props }) => props.max,
+      progress: ({ props }) => props.progress,
+      value: ({ props }) => props.value
     }
   }
 }
 
 export const ProgressCircleWithIcon = {
   extend: ProgressCircle,
-  Icon: { props: { icon: 'arrowUp' } },
-  props: {
-    Icon: {
-      position: 'absolute',
-      top: '50%',
-      left: '50%',
-      transform: 'translate(-50%, -50%)',
-      zIndex: '3',
-      fontSize: 'E'
-    }
+  Icon: {
+    name: 'arrowUp',
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    zIndex: '3',
+    fontSize: 'E'
   }
 }
 
 export const ProgressCircleWithUnitValue = {
   extend: ProgressCircle,
-
-  progress: {},
-  UnitValue: {},
-
-  props: {
-    UnitValue: {
-      position: 'absolute',
-      top: '50%',
-      left: '50%',
-      transform: 'translate(-50%, -50%)',
-      zIndex: '3',
-      fontSize: 'Z'
-    }
+  Progress: {},
+  UnitValue: {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    zIndex: '3',
+    fontSize: 'Z'
   }
 }
 
 export const ProgressCircleWithSideUnitValue = {
   extend: Flex,
-
-  ProgressCircle: {},
-  UnitValue: {},
-
   props: {
     align: 'center flex-start',
     boxSize: 'fit-content fit-content',
-    gap: 'X+V',
-    ProgressCircle: {
-      boxSize: 'A A',
-      ':after': {
-        width: 'calc(100% - 5px)',
-        height: 'calc(100% - 5px)'
-      }
+    gap: 'X+V'
+  },
+
+  ProgressCircle: {
+    boxSize: 'A A',
+    ':after': {
+      width: 'calc(100% - 5px)',
+      height: 'calc(100% - 5px)'
     }
-  }
+  },
+  UnitValue: {}
 }
