+++ uikit/domql/Upload/UploadButton.js
@@ -1,12 +1,9 @@
 'use strict'
 
-import { Input } from '@symbo.ls/input'
 import { Button } from '@symbo.ls/button'
 
 export const UploadButton = {
   extend: Button,
-
-  upload: { extend: Input },
   props: {
     text: 'Choose file',
     position: 'relative',
@@ -16,18 +13,19 @@ export const UploadButton = {
     overflow: 'hidden',
     fontWeight: '500',
     background: 'transparent',
-    cursor: 'pointer',
-    upload: {
-      type: 'file',
-      inset: '0 0 0 0',
-      border: '2px solid red',
-      position: 'absolute',
-      boxSize: '100% 100%',
-      top: '0',
-      left: '0',
-      opacity: '0',
-      cursor: 'pointer'
-    }
+    cursor: 'pointer'
+  },
+
+  Input: {
+    type: 'file',
+    inset: '0 0 0 0',
+    border: '2px solid red',
+    position: 'absolute',
+    boxSize: '100% 100%',
+    top: '0',
+    left: '0',
+    opacity: '0',
+    cursor: 'pointer'
   }
 }
 
@@ -39,8 +37,11 @@ export const UploadButtonWithIcon = {
     background: '#0474F2',
     padding: 'A B',
     text: 'Browse files',
-    round: 'A',
-    icon: {
+    round: 'A'
+  },
+
+  Icon: {
+    props: {
       name: 'upload',
       fontSize: 'B'
     }
