+++ uikit/domql/Upload/UploadModal.js
@@ -1,145 +1,147 @@
 'use strict'
 
-import { Flex } from '@symbo.ls/atoms'
-import { Link } from '@symbo.ls/link'
-import { IconText } from '@symbo.ls/icon'
-import { Dialog } from '@symbo.ls/dialog'
-import { UploadLabel } from './UploadLabel'
+'use strict'
 
-export const UploadModalFooter = {
-  extend: Flex,
-  props: {
-    margin: 'X2 -X2 X2 W',
-    align: 'center space-between'
+import { Modal } from '@symbo.ls/modal'
+import { UploadLabel } from './UploadLabel'
+import { UploadButtonWithIcon } from './UploadButton'
+
+export const UploadFileModal = {
+  extend: Modal,
+
+  header: {
+    title: { h5: { props: { text: 'File Upload' } } },
+    paragraph: null
   },
 
-  IconText: {
-    extend: [IconText, Link],
-    props: {
-      icon: 'info',
-      color: 'gray .65 120',
-      fontWeight: 'regular',
-      target: '_blank',
-      gap: 'Y1',
-      fontSize: 'Z2',
-      textTransform: 'capitalize',
-      text: 'support'
+  content: {
+    extend: UploadLabel,
+    Icon: {},
+    paragraph: {
+      title: {
+        h5: {},
+        UploadButton: null
+      }
     }
   },
 
-  DialogFooter: {
-    props: {
-      padding: '0',
-      gap: 'Y',
-      childProps: {
-        round: 'A'
-      }
-    },
+  footer: {
+    extend: UploadButtonWithIcon,
+    icon: null
+  },
 
-    cancel: {
-      text: 'cancel'
+  props: {
+    gap: 'A',
+    header: { padding: '- - - Z' },
+    content: {
+      padding: 'C E',
+      border: '1px dashed #57575C'
     },
-    ok: {
-      props: {
-        theme: 'secondary'
-      },
-      text: 'Attach File'
+    footer: {
+      text: 'Choose file',
+      boxSize: 'fit-content 100%'
     }
   }
 }
 
-export const UploadModalLabel = {
-  extend: UploadLabel,
-  props: {
-    margin: 'X -X',
-    cursor: 'pointer',
-    padding: 'C1',
-    gap: 'X2',
-    round: 'Z1',
-    userSelect: 'none',
-    border: '1px, dashed, gray 1 25',
-    background: 'gray 1 6',
-    transition: 'A defaultBezier',
-    transitionProperty: 'background',
-    ':hover': {
-      background: 'gray 1 5'
+export const UploadFileModal2 = {
+  extend: Modal,
+
+  header: {
+    title: { h5: { props: { text: 'File Upload' } } },
+    paragraph: null
+  },
+
+  content: {
+    UploadLabel: {}
+  },
+
+  footer: {
+    IconText: {
+      props: {
+        icon: { name: 'info' },
+        text: 'Support'
+      }
+    },
+    CancenConfirmButtons: {
+      ...[
+        {},
+        { text: 'Attach file' }
+      ]
     }
   },
-  Input: {},
-  Icon: {
-    fontSize: 'D',
-    color: 'gray 1 150'
-  },
-  P_2: {
-    extend: Flex,
-    props: {
-      margin: 'W -',
-      gap: 'X2',
-      flexWrap: 'wrap',
-      textAlign: 'center',
-      align: 'center center'
-    },
-    childExtend: {
-      tag: 'span',
-      props: {
-        whiteSpace: 'nowrap',
-        '&:nth-child(1)': {
-          fontSize: 'A',
-          fontWeight: '400',
-          color: '#CFCFD1'
-        },
-        '&:nth-child(2)': {
-          fontSize: 'A',
-          fontWeight: '500',
-          theme: 'primary @dark .color-only',
-          cursor: 'pointer'
+
+  props: {
+    gap: 'A',
+    header: { padding: '- - - Z' },
+    content: { flow: 'column' },
+    footer: {
+      align: 'center space-between',
+      padding: '- - - Z',
+      IconText: { gap: 'Z' },
+      CancenConfirmButtons: {
+        childProps: {
+          ':first-child': { background: '#1C1C1F' }
         }
       }
-    },
-    ...[
-      { text: 'Drag & drop your files here or' },
-      { text: 'Choose file' }
-    ]
-  },
-  P: {
-    text: '50 MB max file size',
-    Span: null
+    }
   }
 }
 
-export const UploadModal = {
-  extend: [Flex, Dialog],
+export const UploadFileModal3 = {
+  extend: UploadFileModal2,
+  header: {},
+  content: {
+    UploadLabel: {},
+    UploadingProcess: {}
+  },
+  footer: {},
 
   props: {
-    flow: 'column',
-    padding: 'Z A',
-    width: '100%',
-    maxWidth: 'H+C',
-    gap: 'X2',
-    round: 'Z1'
+    content: {
+      gap: 'A',
+      UploadingProcess: {
+        boxSize: 'fit-content 100%',
+        content: {
+          flex: '1',
+          ProgressLine: { minWidth: '100%' }
+        }
+      }
+    }
+  }
+}
+
+export const UploadFileModal4 = {
+  extend: UploadFileModal2,
+
+  header: {},
+
+  content: {
+    UploadLabel: null,
+    UploadLabel2: {}
   },
 
-  Flex_header: {
-    props: {
-      padding: '0 X',
-      align: 'center space-between'
-    },
-    H6: {
-      text: 'file upload',
-      fontSize: 'A1',
-      textTransform: 'capitalize'
-    },
-    SquareButton: {
-      margin: '- -Y2 - -',
-      color: 'gray 1 105',
-      theme: 'tertiary',
-      icon: 'x',
-      ':hover': { theme: 'secondary' },
-      ':active': { theme: 'secondary @dark :active' }
+  footer: {
+    IconText: null,
+    CancenConfirmButtons: {
+      minWidth: '100%',
+      childProps: { flex: '1' }
     }
   },
+  props: { footer: { padding: '0' } }
+}
 
-  UploadModalLabel: {},
-
-  UploadModalFooter: {}
+export const UploadFileModal5 = {
+  extend: UploadFileModal4,
+  header: {},
+  content: {
+    UploadLabel2: { border: '1px dashed #3F3F43' },
+    UploadingProcess3: {
+      background: '#040404',
+      boxSize: 'fit-content 100%',
+      border: '1px solid #3F3F43',
+      content: { flex: '1' }
+    }
+  },
+  props: { content: { gap: 'A' } }
 }
