--- uikit/domql/UploadingProcess/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@symbo.ls/uploadingprocess",
-  "version": "2.11.132",
-  "main": "index.js",
-  "license": "MIT",
-  "gitHead": "ce24ae69aa4be9b21beea8118953d112d7e62bc3",
-  "source": "index.js",
-  "dependencies": {
-    "@symbo.ls/atoms": "latest",
-    "@symbo.ls/avatar": "latest",
-    "@symbo.ls/button": "latest",
-    "@symbo.ls/icon": "latest"
-  }
-}
