+++ uikit/domql/index.js
@@ -9,14 +9,10 @@ export * from '@symbo.ls/check'
 export * from '@symbo.ls/titleparagraph'
 export * from '@symbo.ls/field'
 export * from '@symbo.ls/search'
-export * from '@symbo.ls/uploadbutton'
-export * from '@symbo.ls/progress'
-export * from '@symbo.ls/uploadlabel'
 export * from '@symbo.ls/unitvalue'
-export * from '@symbo.ls/uploadingprocess'
 export * from '@symbo.ls/modal'
-export * from '@symbo.ls/uploadfilemodal'
 export * from '@symbo.ls/list'
+export * from '@symbo.ls/progress'
 
 export * from '@symbo.ls/avatar'
 export * from '@symbo.ls/button'
@@ -45,6 +41,5 @@ export * from '@symbo.ls/chat'
 export * from '@symbo.ls/card'
 export * from '@symbo.ls/infoset'
 export * from '@symbo.ls/accessories'
-export * from '@symbo.ls/icontextwithnotification'
 // export * from '@symbo.ls/textcomponents'
 export * from '@symbo.ls/slidetabs'
