+++ uikit/domql/package.json
@@ -20,13 +20,9 @@
     "@symbo.ls/titleparagraph": "latest",
     "@symbo.ls/field": "latest",
     "@symbo.ls/search": "latest",
-    "@symbo.ls/uploadbutton": "latest",
     "@symbo.ls/progress": "latest",
-    "@symbo.ls/uploadlabel": "latest",
     "@symbo.ls/unitvalue": "latest",
-    "@symbo.ls/uploadingprocess": "latest",
     "@symbo.ls/modal": "latest",
-    "@symbo.ls/uploadfilemodal": "latest",
     "@symbo.ls/list": "latest",
     "@domql/utils": "latest",
     "@symbo.ls/atoms": "latest",
@@ -59,7 +55,6 @@
     "@symbo.ls/video": "latest",
     "@symbo.ls/infoset": "latest",
     "@symbo.ls/accessories": "latest",
-    "@symbo.ls/icontextwithnotification": "latest",
     "@symbo.ls/textcomponents": "latest",
     "@symbo.ls/slidetabs": "latest"
